ID:4170 Fix MSOnline provider not restoring the state
The generated secret can now be used to get the state back. The secret
can only be used once to prevent replays.

Error messages caused by 'error' url-params have priority over messages
caused by invalid secrets (because if we have an error, we my not have
a state).

Generate a front-end error message, if the logged in user's domain is
invalid.